[SlicerTrack] Update name of virtual folder for 2D images param #27
We would like to make a change to implement the transforms file with the parameter node. However, we would also like to use a virtual folder to store the transforms. The issue is that we already have a param named `VirtualFolder`, which can be a bit misleading regarding which folder it is pointing to.

This PR changes the `VirtualFolder` param to `VirtualFolder2DImages`.

Former-commit-id: 2989f88e5dbe884ecd5d0f716c1ab5a1f6eb4827