[SlicerTrack] Move warning out of function loadImagesIntoVirtualFolder()
The purpose of this PR is to move the warning that happens if there was an issue when loading the 2D images out of the function `loadImagesIntoVirtualFolder()`. The warning will now be thrown if the return from this function is `None`. This allows us to be consistent with the behaviour of the Transforms file validation function `validateTransformsInput()`.

Former-commit-id: 49b7db1409311eaef5e91ca2dcd8dbd4464637ad