[SlicerTrack][Logic] Implement 'Play' button (#31)
This PR enables the use of the 'Play' button to begin the playback of our 2D images and visualize the alignment of the 3D segmentation against these images according to the transformation data.

This is only a primitive/simple implementation of the 'Play' button which will just play the visual alignment of all the images without any finer control or the ability to stop. Future PRs will be made to implement those features.

Former-commit-id: eca36ab27f3170065a737cb2d73dd3f1578c433d