[SlicerTrack][Data] Delete extra data that is no longer needed (#42)
This PR removes unneeded data files within our data folder.

In previous versions of SlicerTrack, we were exploring an approach where we "stack" the 2D images together and work with them in this way. We no longer need to stack the images and are able to step through them one by one.

We remove `Segmentation_nrrd.nrrd` because for our use-case it is functionally the same as `Segmentation.mha`.

Former-commit-id: aca85b08d4507af32df79cfe7802d676bce80723