[SlicerTrack][Logic] Use label map of 3D segmentation as a better mask to overlay on the 2D images (#45)
The purpose of this PR is to make a label map of our 3D segmentation and to use this as the "label layer" in the slice view. This label map acts as a better mask of the 3D segmentation to overlay on the 2D images. This label map only covers the exact ROI area (while the 3D volume would darken the surrounding) and also provides some useful features (such as being able to view the mask either as a filled region vs. an outline).