[SlicerTrack] Fix module reload bug (#46)
When the module is reloaded (i.e. the module is destroyed and then recreated while being provided the same parameter node) an error would appear that would indicate the `totalImages` variable was None. This was because after the module was re-created, `totalImages` was given it's default value (None) and it was never set to the number of images within the 2D images virtual folder.

This PR fixes the bug by setting the `totalImages` var in the `setParameterNode()` function, which is called after the module is created. If it finds the 2D images folder in the provided parameter node, it will set the `totalImages` variable accordingly.

Former-commit-id: 1f5ba1ae2a6fc6ae93e7394df966d98f4d401add