[SlicerTrack][Logic] Display "Pre Alignment" and "Post Alignment" text in slice view corner (#47)
The purpose of this PR is to add text in the slice view corner that will make it easier for the user to recognize which image is being aligned to. It will display the text "Pre Alignment" when the image and the overlay has been visualized, but no transformation has been applied. And then it will display "Post Alignment" in the corner when the transformation is applied to align the 3D segmentation to the 2D image.

Former-commit-id: 2121bfe605ee35d2d3ddb7b3a1cee4172d6efb63