[SlicerTrack] Implement loading/progress bar for images/transformation data importation (#50)
When using the SlicerTrack extension loading the 2D images and the transformation data from the transforms file can take a bit of time, and the GUI is frozen while this is happening. This slowness is exacerbated on slower systems.

This PR implements a progress bar dialog which is shown at these importation steps. The bar allows the user to anticipate how much longer it will take to load the data into 3D Slicer.

Former-commit-id: f7a5e5929a513d272b1a18ec1ae10efc114c3492