[SlicerTrack][Logic] Change how we convert the transformation data so that it can be used in 3D Sllicer (#54)
The purpose of this PR is to update how we convert the transformation data within SlicerTrack. Previously, we assumed that the conversion information was related to the direction matrix within the metadata of the 2D images (which can be different for a sagittal vs. coronal image). 

It seems that the correct approach is to simply apply a LPS to RAS conversion, since the transformation data was generated outside of 3D Slicer with DICOM images that use the LPS basis, whereas 3D Slicer uses the RAS basis.

Former-commit-id: 8d75394fdb16e063d6ed72cf2528c122c139dde8