[SlicerTrack][Logic] Remove the pre-alignment step (#55)
After some discussion with Dr. Lugez post-IMNO conference we realized we do not want to have the 'pre-alignment' step where the 'default' alignment/overlay of the 3D segmentation over the 2D image is shown. Primarily because it is misleading to the user, but also because it is an unnecessary spec.

Former-commit-id: 989139aaa8d430108995d3fdde705add0186280a