[GUI] Implement playback speed multiplier control (#58)
This PR replaces our FPS input box with a playback speed input box. The motivation for this PR was to simplify how the user makes the playback slower or faster, and to not have them concerned with the precise amount of frames per second (FPS). It is also difficult to guarantee that the playback will be playing at exactly a particular FPS value.