[Logic][Parameter Node] Integrate Sequence and SequenceBrowser nodes into backend (#69)
The purpose of this PR is to implement the usage of `Sequence` and `SequenceBrowser` nodes in the backend of SlicerTrack.

We also hookup the Sequence Browser operation with the media player GUI buttons. This simplifies our code in a lot of places.

This should resolve #62 

Note: We no longer visualize the 2D images and 3D segmentation in the 3D view due to slowness during playback at higher speeds. I have left the code for this functionality commented out.

For Testing NOTE: This change may crash 3D Slicer for you when using it alongside the Sequences module, if you are not using the latest preview release.
Former-commit-id: e647d1c43b775f3fae639d1d46e52df6ac09ae97