[GUI] Add Image information labels, [Logic] Fix Image spinbox and Sequence Slider behaviour (#90)
This PR allows the user to change the current image of the sequence when it is not being played. This can be done by modifying the value of either the Sequence Slider or the Frame Input Box. This addresses issues #70 and #78.
Additionally, this PR adds labels to the "Sequence" section of this extension by displaying "Image _ of _" to the right of the sequence slider. This addresses issue #71.

This PR intends to close #70, close #71, and close #78

Tested on MacOS (Version 5.3.0 - Preview Release)

Former-commit-id: 017c74db417ec763627c50004b8ee0c5c440d69e