Preserve slice views and add minor changes (#95)
This PR preserves all slice views, so when the next images in the sequence appear in a different slice view, the last image using the other slice view also gets displayed. If no image was previously displayed in that slice view, it stays empty. Since we are using Sequence nodes, a node for each slice view is created (`self.redBackground` `self.greenBackground` `self.yellowBackground`) to keep track of the latest image in each slice view from the sequence. The `setupSliceViews()` method is renamed to `visualize()`, to navigate between Sequence nodes.

Additionally, the following minor changes are included:

- Replace "time-series Images" with "cine Images"
- Accept all .mha files with no naming restriction when loading a folder for 2D Cine images
- Group all cine images into one folder
- Fix issues when interacting with `onIncrement()`, `onDecrement()` and `onStopButton()`methods

Tested on MacOS (Version 5.3.0 - Preview Release)